# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Serviço Backend",
        "description": "API RESTful para gerenciamento de serviços comunitários, incluindo pessoas, perfis, missões, notícias e doações.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/candidaturas-missao": {
            "get": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Lista todas as candidaturas",
                "operationId": "c2f8aacfbb0b3270263550a156a8c78c",
                "responses": {
                    "200": {
                        "description": "Lista de candidaturas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Cria uma nova candidatura",
                "operationId": "23773ea8e969c4319686abd2e14a4e4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "missao_id",
                                    "voluntario_id"
                                ],
                                "properties": {
                                    "missao_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "voluntario_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidatura criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/missoes/{missaoId}/candidaturas": {
            "get": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Lista candidaturas de uma missão",
                "operationId": "4d1f5bf29f2e594af97b256e102b9678",
                "parameters": [
                    {
                        "name": "missaoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de candidaturas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/voluntarios/{voluntarioId}/candidaturas": {
            "get": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Lista candidaturas de um voluntário",
                "operationId": "5fc4091c105f1805db2f94a3e209a9af",
                "parameters": [
                    {
                        "name": "voluntarioId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de candidaturas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidaturas-missao/{id}": {
            "get": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Busca candidatura por ID",
                "operationId": "957478c85efd0a1e8a8c8fe6ceace401",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidatura encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidatura não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Exclui uma candidatura",
                "operationId": "c3d02f93e39d97ed2b7b4af47d847d40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Candidatura excluída"
                    },
                    "404": {
                        "description": "Candidatura não encontrada"
                    }
                }
            }
        },
        "/api/candidaturas-missao/{id}/aprovar": {
            "patch": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Aprova uma candidatura",
                "operationId": "f0cd56ab4abc79db836bfe070f4ddd47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidatura aprovada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidatura não encontrada"
                    },
                    "409": {
                        "description": "Missão não possui mais vagas disponíveis"
                    }
                }
            }
        },
        "/api/candidaturas-missao/{id}/rejeitar": {
            "patch": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Rejeita uma candidatura",
                "operationId": "fd12535a9e08c396eb4e9c0ab3ae8d22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "obs": {
                                        "type": "string",
                                        "example": "Motivo da rejeição"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidatura rejeitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidatura não encontrada"
                    }
                }
            }
        },
        "/api/candidaturas-missao/{id}/concluir": {
            "patch": {
                "tags": [
                    "Candidaturas"
                ],
                "summary": "Conclui uma candidatura com avaliação",
                "operationId": "6f1d26cd96761f808f00b700282eafb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "avaliacao": {
                                        "description": "Avaliação de 1 a 5",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "obs_avaliacao": {
                                        "type": "string",
                                        "example": "Observações da avaliação"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidatura concluída",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidatura não encontrada"
                    }
                }
            }
        },
...
    },
...
}